refactor(appstate): split VoiceFlowKit speech integration into AppState+VoiceFlowKit.swift
Following the same same-type extension pattern VoiceFlow used in
grapeot/voiceflow#40, peel the speech integration code off the main
AppState file. Net -97 lines from AppState.swift.

Move to `AppState+VoiceFlowKit.swift`:
- `makeVoiceFlowClient()` — builds VoiceFlowConfig + VoiceFlowClient
- `transcribeAudio(audioFileURL:onPartialTranscript:)` — one-shot bulk
- `startRealtimeSpeechSession()` — chat composer realtime entry
- `testAIBuilderConnection()` — Settings-side connection test with
  typed VoiceFlowError pattern-matching

Demoted three `private static` members to internal (default) so the
extension can use them: `logger`, `aiBuilderLastOKSignatureKey`,
`aiBuilderLastOKTestedAtKey`, and `aiBuilderSignature(baseURL:token:)`.
Still module-internal, no API surface change.

Tests: TEST SUCCEEDED (full OpenCodeClientTests suite passes).

This is the first of several planned AppState splits — SSE / Sessions
/ FileTree extensions are larger and more interdependent, will be done
in follow-up PRs.
